Revert "refactor: 홈 내용을 Home 컴포넌트로 압축"
This reverts commit 74b9be111aad2a0e4cc696f47f8b19d84d5fa953.
