+++ bubbletea03-my-app/src/App.js
@@ -1,23 +1,24 @@
 import './App.css';
-
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Home from './component/home/Home';
+import Footer from './component/Footer';
+import Header from './component/Header';
+import ItemCard from './component/ItemCard.js';
+import PlusButton from './component/PlusButton';
 
 function App() {
   return (
     <div className="App">
-      <BrowserRouter>
-				<Routes>
-					<Route path="*" element={<div> Not Found </div>} />
-
-					<Route path="/" element={
-            <div>
-              <Home/>
-            </div>
-          }/>
-					<Route path="/product-page" element={<div> DD </div>} />
-				</Routes>
-			</BrowserRouter>
+      <i class="fa-solid fa-bars"></i>
+      <Header/>
+      <ItemCard/>
+      <ItemCard/>
+      <ItemCard/>
+      <ItemCard/>
+      <ItemCard/>
+      <ItemCard/>
+      <ItemCard/>
+      <ItemCard/>
+      <PlusButton/>
+      <Footer/>
     </div>
   );
 }
