--- bubbletea03-my-app/src/component/home/Home.js
@@ -1,24 +1,0 @@
-import Footer from './Footer.js';
-import Header from './Header.js';
-import ItemCard from './ItemCard.js';
-import PlusButton from './PlusButton.js';
-
-function Home() {
-    return (
-        <div className="header">
-            <Header/>
-            <PlusButton/>
-            <Footer/>
-            <ItemCard/>
-            <ItemCard/>
-            <ItemCard/>
-            <ItemCard/>
-            <ItemCard/>
-            <ItemCard/>
-            <ItemCard/>
-            <ItemCard/>
-        </div>
-    );
-}
-
-export default Home;
